+++ inc/FixedVector.hh
@@ -22,7 +22,7 @@ public:
   T& operator[](std::size_t idx)
   {
     if (idx >= vec.size())
-      throw std::out_of_range("Index out of bounds. Vec size: %d", vec.size());
+      throw std::out_of_range("Index out of bounds. Vec size: " + std::to_string(vec.size()));
     return vec[vec.size() - idx - 1];
   }
