+++ unittest/RunUnitTest.md
@@ -1,5 +1,5 @@
 To run the unit tests and confirm functionality, navigate to this dir and run:  
-g++ -std=c++17 -o test test.cc
+g++ -std=c++17 -o unittest UnitTest.cc
 
 This will generate an exe to run the unit tests so then run
-./test
+./unittest
