+++ unittest/UnitTest.cc
@@ -1,5 +1,6 @@
 #include <iostream>
 #include "../inc/FixedVectorMath.hh"
+#include "../inc/FixedVector.hh"
 
 template <typename T>
 void Print1DVec(const FixedVector<T>& v) {
@@ -18,8 +19,12 @@ int main() {
 
     size_t size = 4;
     FixedVector<float> intVector(size, 0);
+    for (int i = 0; i < intVector.size(); i++) {
+        intVector[i] = 3;
+    }
     Print1DVec(intVector);
     FixedVectorMath::normalize(intVector);
+    std::cout << "TEST 1: Output" << std::endl;
     Print1DVec(intVector);
 
     std::cout << "Unit Testing Completed\n";
